update Transaction Modal
